gh-266: Moved out all Iterable classes into iterable module
