+++ core/src/main/java/uk/gov/gchq/koryphe/iterable/ChainedIterator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.koryphe.commonutil.iterable;
+package uk.gov.gchq.koryphe.iterable;
 
 import uk.gov.gchq.koryphe.util.CloseableUtil;
